add const and let guide
With some wording and spelling changes, a guide to use `const` and `let` rather than `var` has been added. This is likely to be one of the most controversial aspects of this guide, but from all the evidence and experience I can gather, `var` should be avoided in favor of it's new successors.